+++ Sources/ZcashLightClientKit/Rust/ZcashKeyDerivationBackendWelding.swift
@@ -70,4 +70,38 @@ protocol ZcashKeyDerivationBackendWelding {
     ///     - `rustGetTransparentReceiverInvalidAddress` if failed to get transparent receiver for unified address.
     ///     - `rustGetTransparentReceiverInvalidReceiver` if generated transparent receiver is invalid.
     func getTransparentReceiver(for uAddr: UnifiedAddress) throws -> TransparentAddress
+
+    /// Derives and returns a ZIP 32 Arbitrary Key from the given seed at the "wallet level", i.e.
+    /// directly from the seed with no ZIP 32 path applied.
+    ///
+    /// The resulting key will be the same across all networks (Zcash mainnet, Zcash testnet,
+    /// OtherCoin mainnet, and so on). You can think of it as a context-specific seed fingerprint
+    /// that can be used as (static) key material.
+    ///
+    /// - Parameter contextString: a globally-unique non-empty sequence of at most 252 bytes that identifies the desired context.
+    /// - Parameter seed: `[Uint8]` seed bytes
+    /// - Parameter accountNumber: `Int` with the account number
+    /// - Throws:
+    ///     - `derivationToolInvalidAccount` if the `accountIndex` is invalid.
+    ///     - some `ZcashError.rust*` error if the derivation fails.
+    /// - Returns a `[Uint8]`
+    static func deriveArbitraryWalletKey(
+        contextString: [UInt8],
+        from seed: [UInt8]
+    ) throws -> [UInt8]
+
+    /// Derives and returns a ZIP 32 Arbitrary Key from the given seed at the account level.
+    ///
+    /// - Parameter contextString: a globally-unique non-empty sequence of at most 252 bytes that identifies the desired context.
+    /// - Parameter seed: `[Uint8]` seed bytes
+    /// - Parameter accountNumber: `Int32` with the account number
+    /// - Throws:
+    ///     - `derivationToolInvalidAccount` if the `accountIndex` is invalid.
+    ///     - some `ZcashError.rust*` error if the derivation fails.
+    /// - Returns a `[Uint8]`
+    func deriveArbitraryAccountKey(
+        contextString: [UInt8],
+        from seed: [UInt8],
+        accountIndex: Int32
+    ) throws -> [UInt8]
 }