+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this library adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # Unreleased
 
+## Added
+- `DerivationTool.deriveArbitraryWalletKey`
+- `DerivationTool.deriveArbitraryAccountKey`
+
 # 2.2.6 - 2024-10-22
 
 ## Fixed